+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [0ver](https://0ver.org).
 
 ## [Unreleased]
 
+### Added
+
+- A couple commands to enable/disable the operations on the workspace from the CLI
+
 ## [0.0.7] - 2020-04-09
 
 ### Changed